Optimize error handling & provide fast path for non-blocking send
runtime.LockOSThread() has a cost associated with it, as well as each
cgo call.
Also provide a fastpath for non-blocking mode: try to send first and
only start the epoller when send fails, this reduces a cgo call in the
happy case (most likely the default case).
